requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==48.0.0
distro==1.9.0
fastapi==0.136.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.196.0
google-auth==2.52.0
google-auth-httplib2==0.4.0
google-genai==2.0.0
googleapis-common-protos==1.75.0
groq==1.2.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.13
proto-plus==1.28.0
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
pyparsing==3.3.2
python-dotenv==1.2.2
python-telegram-bot==22.7
requests==2.33.1
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.7.0
uvicorn==0.46.0
websockets==16.0
